+++ src/components/signal-detail-sidebar/index.js
@@ -10,13 +10,27 @@ import {
   navigation,
 } from 'nr1';
 
-import { SIGNAL_TYPES } from '../../constants';
+import { SIGNAL_TYPES, STATUSES } from '../../constants';
 
 import Incidents from './incidents';
 import GoldenMetrics from './golden-metrics';
 import { AppContext } from '../../contexts';
 
-const SignalDetailSidebar = ({ guid, name, type }) => {
+import typesList from '../../../nerdlets/signal-selection/types.json';
+
+const NO_ENTITY_TYPE = '(unknown entity type)';
+
+const entityTypeFromGuid = (guid) => {
+  if (!guid) return NO_ENTITY_TYPE;
+  const [, domain, type] = atob(guid)?.split('|') || [];
+  if (!domain || !type) return NO_ENTITY_TYPE;
+  return (
+    typesList.find((t) => t.domain === domain && t.type === type)
+      ?.displayName || NO_ENTITY_TYPE
+  );
+};
+
+const SignalDetailSidebar = ({ guid, name, type, status }) => {
   const { account = {}, accounts = [] } = useContext(AppContext);
   const [conditionId, setConditionId] = useState();
   const [hasAccessToEntity, setHasAccessToEntity] = useState(false);
@@ -54,9 +68,11 @@ const SignalDetailSidebar = ({ guid, name, type }) => {
             </HeadingText>
             <HeadingText type={HeadingText.TYPE.HEADING_3}>{name}</HeadingText>
             <HeadingText type={HeadingText.TYPE.HEADING_5}>
-              {signalAccount
-                ? `${signalAccount?.name} | ${signalAccount?.id}`
-                : ''}
+              {`${signalAccount?.name || '(unknown account)'} | ${
+                type === SIGNAL_TYPES.ENTITY
+                  ? entityTypeFromGuid(guid)
+                  : 'Alert condition'
+              }`}
             </HeadingText>
             {hasAccessToEntity ? (
               <Link
@@ -76,6 +92,7 @@ const SignalDetailSidebar = ({ guid, name, type }) => {
             type={type}
             conditionId={conditionId}
             accountId={signalAccount.id}
+            status={status}
           />
           {type === SIGNAL_TYPES.ENTITY ? <GoldenMetrics guid={guid} /> : null}
         </>
@@ -93,6 +110,7 @@ SignalDetailSidebar.propTypes = {
   guid: PropTypes.string,
   name: PropTypes.string,
   type: PropTypes.oneOf(Object.values(SIGNAL_TYPES)),
+  status: PropTypes.oneOf(Object.values(STATUSES)),
 };
 
 export default SignalDetailSidebar;